:art: Refactor options to a more split out structure
