+++ src/configurator/options/BaseOptions.js
@@ -1,0 +1,5 @@
+export default class BaseOptions {
+    strength() {
+        return { min: 0, max: 10 };
+    }
+}
